// web/lib/api/generated/model/unit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export interface Unit {
  readonly id: string;
  /** @maxLength 10 */
  code: string;
  /** @maxLength 60 */
  name: string;
  allows_decimal?: boolean;
  /** @maxLength 10 */
  uqc: string;
  is_active?: boolean;
}